requirements.txt
defusedxml==0.5.0
gdata==2.0.18
iso8601==0.1.11
jira==1.0.10
MySQL-python==1.2.5
oauthlib==1.0.3
ordereddict==1.1
pbr==3.0.1
python-decouple==3.0
requests==2.9.1
requests-oauthlib==0.6.1
requests-toolbelt==0.6.0
six==1.10.0
tlslite==0.4.9
